Run success

+++ STRM/src/main/cpp/strm/include/strm/impl/models/TfLiteYoloV5ClassifierDSP.hpp
@@ -35,8 +35,8 @@ class TfLiteYoloV5ClassifierDSP : public Classifier {
 
   float inputScale;
   float outputScale;
-  int* input; // 1 x inputSize.height x inputSize.width x 3
-  int* outputs; // 1 x outputSize x 85 (boxes, maxConf, confidences)
+  uint8_t* input; // 1 x inputSize.height x inputSize.width x 3
+  uint8_t* outputs; // 1 x outputSize x 85 (boxes, maxConf, confidences)
 };
 
 } // namespace rm
